requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
urllib3
flask==3.1.2
flask-cors==6.0.1